Split TestInput and TestInputs implementation
